+++ crates/holochain/src/conductor/api/api_external/admin_interface.rs
@@ -307,6 +307,10 @@ impl AdminInterfaceApi for RealAdminInterfaceApi {
                     .await?;
                 Ok(AdminResponse::FullStateDumped(state))
             }
+            DumpNetworkMetrics { dna_hash } => {
+                let dump = self.conductor_handle.dump_network_metrics(dna_hash).await?;
+                Ok(AdminResponse::NetworkMetricsDumped(dump))
+            }
             AddAgentInfo { agent_infos } => {
                 self.conductor_handle.add_agent_infos(agent_infos).await?;
                 Ok(AdminResponse::AgentInfoAdded)